Add a Style class
